Rename: step3 진행을 위한 문서 이동
