+++ src/main/java/simpleCalculator/view/CalculatorView.java
@@ -1,4 +1,4 @@
-package view;
+package simpleCalculator.view;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Scanner;
